readify_eval/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy>=2.0.25
pydantic>=2.9.2
pydantic-settings==2.5.0
PyYAML==6.0.1
python-dotenv==1.0.0
alembic==1.12.1
pymysql==1.1.0
cryptography>=41.0.0
asyncpg==0.29.0
python-multipart==0.0.6
httpx>=0.25.0
sentence-transformers>=2.2.0
numpy>=1.24.0
torch>=2.0.0
ragas>=0.2.0
datasets>=3.0.0
pyarrow>=17.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
sacrebleu>=2.0.0
rouge_score>=0.1.0
nacos-sdk-python>=3.0.2